requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
yfinance
python-dotenv
psycopg2-binary